requirements.txt
# Core dependencies
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.8.0

# Audio processing
pyaudio==0.2.11
numpy==1.24.3
scipy==1.11.2

# Machine Learning - Transcription
openai-whisper==20230918
torch==2.0.1
torchaudio==2.0.2

# Optional - Speaker Diarization (requires additional setup)
# pyannote.audio==3.1.1

# Utilities
pathlib2==2.3.7
python-dateutil==2.8.2
